requirements.txt
PyGithub
theblues